pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "cachepilot"
version = "0.1.0"
description = "Real-time AI memory orchestrator for multi-model GPU serving"
authors = [{ name = "Aryan Putta", email = "aryansputta@gmail.com" }]
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
    "typer>=0.12",
    "pydantic>=2.0",
    "numpy>=1.26",
    "pandas>=2.0",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.optional-dependencies]
plot = ["matplotlib>=3.8"]
dev  = ["pytest>=8.0", "pytest-cov>=5.0", "ruff>=0.4"]

[project.scripts]
cachepilot = "cachepilot.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts   = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]
